Begin BLEScanner refactor
Creates API for adding BLECentralRequests to be executed in order.
Allows us to more easily modify behavior of BLECentral and consolidates
the logic for a single request into BLEScannerRequest. Also establishes
a queue pattern so that requests are executed only after a response is
received from the previous request to smooth the bluetooth i/o.

TODO : Move the individual request logic to BLETransportManager, and
set it on BLEScanner via addRequest